Move tests to match new locations of common elements
